requirements.txt
appwrite
python-dotenv
fastapi
uvicorn
yt-dlp
python-multipart
pydantic
pathlib